+++ radix-engine/src/model/substates/mod.rs
@@ -1,9 +1,17 @@
 mod component;
 mod key_value_store_entry;
 mod non_fungible;
+mod package;
+mod resource_manager;
+mod substate;
+mod system;
 mod vault;
 
 pub use component::*;
 pub use key_value_store_entry::*;
 pub use non_fungible::*;
+pub use package::*;
+pub use resource_manager::*;
+pub use substate::*;
+pub use system::*;
 pub use vault::*;
