+++ radix-engine/src/model/substates/system.rs
@@ -1,0 +1,6 @@
+use crate::types::*;
+
+#[derive(Debug, Clone, TypeId, Encode, Decode, PartialEq, Eq)]
+pub struct SystemSubstate {
+    pub epoch: u64,
+}
